+++ data/src/main/java/com/umc/data/response/schedule/ScheduleListResponse.kt
@@ -48,9 +48,10 @@ data class ScheduleListResponse (
             )
         }
 
+        /**TODO 서버 DTO에 맞춰 수정 : attendanceRate는 차후 douvle 추천!**/
         fun ScheduleListResponse.toAdminDomain(): AdminSessionCheck {
             return AdminSessionCheck(
-                id = scheduleId.toInt(),
+                id = scheduleId,
                 title = name,
                 date = date,
                 startTime = startTime,
@@ -63,7 +64,6 @@ data class ScheduleListResponse (
                 pendingUsers = emptyList()
             )
         }
-
     }
 }
